SALTO-2628 warn in logs on duplicate type names
