+++ packages/adapter-components/src/config/request.ts
@@ -16,7 +16,9 @@
 import { ElemID, ObjectType, BuiltinTypes, CORE_ANNOTATIONS, FieldDefinition, MapType, ListType, ActionName, createRestriction } from '@salto-io/adapter-api'
 import { createMatchingObjectType } from '@salto-io/adapter-utils'
 import _ from 'lodash'
-import { findDuplicates } from './validation_utils'
+import { collections } from '@salto-io/lowerdash'
+
+const { findDuplicates } = collections.array
 
 export const ARG_PLACEHOLDER_MATCHER = /\{([\w_]+)\}/g
