+++ packages/adapter-components/src/config/transformation.ts
@@ -16,9 +16,10 @@
 import _ from 'lodash'
 import { ElemID, ObjectType, BuiltinTypes, CORE_ANNOTATIONS,
   FieldDefinition, ListType, RestrictionAnnotationType } from '@salto-io/adapter-api'
-import { types, values } from '@salto-io/lowerdash'
-import { findDuplicates } from './validation_utils'
+import { types, values, collections } from '@salto-io/lowerdash'
 import { getConfigWithDefault, TypeConfig, TypeDefaultsConfig } from './shared'
+
+const { findDuplicates } = collections.array
 
 export const DATA_FIELD_ENTIRE_OBJECT = '.'
 export const FIELD_REFERENCE_PREFIX = '&'
